feat(code) Telechargement de fichier depuis le MQTT
[60min][DONE]
Je peux télécharger des fichier entier si une personne l'envoie d'un coup entier mais si il envoie en morceau y a quelque petit problème mais sinon sa marche

+++ P_BitRuisseau-321/IProtocol.cs
@@ -39,7 +39,7 @@ namespace BitRuisseau
         /// <param name="startByte">The first byte you need</param>
         /// <param name="endByte">The last byte you need</param>
         /// <param name="name">The name/ip of the mediatheque</param>
-        public void AskMedia(ISong song, string name, int startByte, int endByte);
+        public void AskMedia(ISong song, string name, long startByte, long endByte);
 
         /// <summary>
         /// Send the media to someone
@@ -48,6 +48,6 @@ namespace BitRuisseau
         /// <param name="startByte">The first byte they need</param>
         /// <param name="endByte">The last byte they need</param>
         /// <param name="name">The name/ip of the mediatheque</param>
-        public void SendMedia(ISong song, string name, int startByte, int endByte);
+        public void SendMedia(ISong song, string name, long startByte, long endByte);
     }
 }
